Rename Provider to hook
