+++ packages/expo-router/src/react-navigation/core/useOnAction.tsx
@@ -16,7 +16,6 @@ import {
   type ChildPreventRemoveListener,
   NavigationBuilderContext,
 } from './NavigationBuilderContext';
-import type { IsRoutePrevented } from './PreventRemoveProvider';
 import type { EventMapCore } from './types';
 import type { NavigationEventEmitter } from './useEventEmitter';
 import {
@@ -25,6 +24,7 @@ import {
   shouldPreventRemove,
   useOnPreventRemove,
 } from './useOnPreventRemove';
+import type { IsRoutePrevented } from './usePreventRemoveState';
 
 type Options<State extends NavigationState> = {
   router: Router<State, NavigationAction>;
