+++ packages/expo-router/src/react-navigation/core/usePreventRemoveState.tsx
